requirements.txt
fastapi
uvicorn
pydantic
requests
streamlit
passlib[bcrypt]
python-jose[cryptography]
python-multipart
